requirements.txt
azure-identity==1.17.1
deltalake==0.18.2